+++ router/relay-router.go
@@ -68,6 +68,7 @@ func SetRelayRouter(router *gin.Engine) {
 		playgroundRouter.POST("/images/generations", controller.PlaygroundImage)
 		playgroundRouter.POST("/video/generations", controller.PlaygroundVideo)
 		playgroundRouter.POST("/3d/generations", controller.Playground3D)
+		playgroundRouter.POST("/audio/speech", controller.PlaygroundAudio)
 	}
 	relayV1Router := router.Group("/v1")
 	relayV1Router.Use(middleware.RouteTag("relay"))